refactor: 잘못된 repository method naming 수정
