+++ src/main/kotlin/com/example/kbocombo/crawler/application/PlayerRenewService.kt
@@ -12,7 +12,7 @@ class PlayerRenewService(
 
     @Transactional
     fun renew(recentPlayers: List<Player>) {
-        val savedPlayers = playerRepository.findAllByRetiredNot()
+        val savedPlayers = playerRepository.findAllByIsRetiredFalse()
             .associateBy { it.webId }
 
         for (recentPlayer in recentPlayers) {
